# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SkillForge AI - Learning Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses/{courseId}/modules:
    get:
      operationId: getModules
      tags: [modules]
      summary: List modules for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [modules]
      summary: Create module
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
  /modules/{moduleId}:
    patch:
      operationId: updateModule
      tags: [modules]
      summary: Update module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModuleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
    delete:
      operationId: deleteModule
      tags: [modules]
      summary: Delete module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Lessons
...
  /progress/{courseId}:
    get:
      operationId: getCourseProgress
      tags: [progress]
      summary: Get progress for a specific course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseProgress"
  /progress/lessons/{lessonId}/complete:
    post:
      operationId: completeLesson
      tags: [progress]
      summary: Mark lesson as complete
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonProgress"

  # Quizzes
  /courses/{courseId}/quizzes:
    get:
      operationId: getQuizzes
      tags: [quizzes]
      summary: List quizzes for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quizzes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quiz"
    post:
      operationId: createQuiz
      tags: [quizzes]
      summary: Create quiz
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
  /quizzes/{quizId}:
    get:
      operationId: getQuiz
      tags: [quizzes]
      summary: Get quiz with questions
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizDetail"
    patch:
      operationId: updateQuiz
      tags: [quizzes]
      summary: Update quiz
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
    delete:
      operationId: deleteQuiz
      tags: [quizzes]
      summary: Delete quiz
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /quizzes/{quizId}/submit:
    post:
      operationId: submitQuiz
      tags: [quizzes]
      summary: Submit quiz answers
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
  /quizzes/{quizId}/attempts:
    get:
      operationId: getQuizAttempts
      tags: [quizzes]
      summary: Get quiz attempts
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Attempts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuizAttempt"
  /quizzes/{quizId}/leaderboard:
    get:
      operationId: getQuizLeaderboard
      tags: [quizzes]
      summary: Get quiz leaderboard
      parameters:
        - name: quizId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  # Assignments
...
  /notifications/mark-read:
    patch:
      operationId: markNotificationsRead
      tags: [notifications]
      summary: Mark notifications as read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarkReadResponse"
...
components:
  schemas:
...
    Module:
      type: object
      required: [id, courseId, title, position]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        position:
          type: integer
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
    ModuleInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        position:
          type: integer
    ModuleUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        position:
          type: integer

    # Lessons
    Lesson:
      type: object
      required: [id, moduleId, title, type, position]
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [video, text, quiz, assignment, resource]
        content:
          type: ["string", "null"]
        videoUrl:
          type: ["string", "null"]
        resourceUrl:
          type: ["string", "null"]
        duration:
          type: ["integer", "null"]
        position:
          type: integer
        isFree:
          type: boolean
        isCompleted:
          type: boolean
...
    CourseProgress:
      type: object
      required: [courseId, userId, progressPercent, completedLessons, totalLessons]
      properties:
        courseId:
          type: integer
        userId:
          type: integer
        progressPercent:
          type: number
        completedLessons:
          type: integer
        totalLessons:
          type: integer
        lastLessonId:
          type: ["integer", "null"]
        completedAt:
          type: ["string", "null"]
    LessonProgress:
      type: object
      required: [lessonId, userId, completedAt]
      properties:
        lessonId:
          type: integer
        userId:
          type: integer
        completedAt:
          type: string
          format: date-time

    # Quizzes
    Quiz:
      type: object
      required: [id, courseId, title, questionCount]
      properties:
        id:
          type: integer
        courseId:
          type: integer
        lessonId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        timeLimit:
          type: ["integer", "null"]
        passingScore:
          type: number
        questionCount:
          type: integer
        attempts:
          type: integer
    QuizDetail:
      allOf:
        - $ref: "#/components/schemas/Quiz"
        - type: object
          properties:
            questions:
              type: array
              items:
                $ref: "#/components/schemas/Question"
    QuizInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        timeLimit:
          type: integer
        passingScore:
          type: number
        lessonId:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuestionInput"
    QuizUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        timeLimit:
          type: integer
        passingScore:
          type: number
    Question:
      type: object
      required: [id, quizId, text, type, position]
      properties:
        id:
          type: integer
        quizId:
          type: integer
        text:
          type: string
        type:
          type: string
          enum: [multiple_choice, true_false, fill_blank]
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
        explanation:
          type: ["string", "null"]
        position:
          type: integer
        points:
          type: integer
    QuestionInput:
      type: object
      required: [text, type]
      properties:
        text:
          type: string
        type:
          type: string
          enum: [multiple_choice, true_false, fill_blank]
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
        correctAnswer:
          type: string
        explanation:
          type: string
        position:
          type: integer
        points:
          type: integer
    QuestionOption:
      type: object
      required: [id, text]
      properties:
        id:
          type: string
        text:
          type: string
        isCorrect:
          type: boolean
    QuizSubmission:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswer"
    QuizAnswer:
      type: object
      required: [questionId, answer]
      properties:
        questionId:
          type: integer
        answer:
          type: string
    QuizResult:
      type: object
      required: [attemptId, score, passed, totalPoints, earnedPoints]
      properties:
        attemptId:
          type: integer
        score:
          type: number
        passed:
          type: boolean
        totalPoints:
          type: integer
        earnedPoints:
          type: integer
        answers:
          type: array
          items:
            $ref: "#/components/schemas/AnswerResult"
    AnswerResult:
      type: object
      required: [questionId, isCorrect, earnedPoints]
      properties:
        questionId:
          type: integer
        isCorrect:
          type: boolean
        earnedPoints:
          type: integer
        explanation:
          type: ["string", "null"]
    QuizAttempt:
      type: object
      required: [id, quizId, userId, score, passed, createdAt]
      properties:
        id:
          type: integer
        quizId:
          type: integer
        userId:
          type: integer
        score:
          type: number
        passed:
          type: boolean
        createdAt:
          type: string
          format: date-time
    LeaderboardEntry:
      type: object
      required: [rank, userId, userName, score, completedAt]
      properties:
        rank:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userAvatar:
          type: ["string", "null"]
        score:
          type: number
        completedAt:
          type: string
          format: date-time

    # Assignments
...
    MarkReadInput:
      type: object
      required: [notificationIds]
      properties:
        notificationIds:
          type: array
          items:
            type: integer
    MarkReadResponse:
      type: object
      required: [updated]
      properties:
        updated:
          type: integer
...